test: sync test suite with current source-of-truth
Aligns tests with changes already landed in production code:

- Exception types: replace RuntimeError / AcquisitionError expectations with
  the centralized hierarchy (LearnerError, OracleError, FeatureExtractionError)
  across cycle, ensemble, sklearn, torch, oracle, and extraction tests.
- Ensemble random_states: [42, 123, 456] -> [42, 123, 356] to match
  ENSEMBLE_SEED_OFFSETS = (0, 81, 314); the prior expectation was a stale
  literal that never matched the derived seeds.
- Parquet schema: integration test now expects Float32 for prediction /
  uncertainty (matches 022-metrics-scale storage downcast).
- EI/PI speedup benchmark: gate behind LEARNM8_BENCHMARK=1 so hardware-
  sensitive 4x targets do not flake on dev hosts (WSL/laptops hit ~1.5x).
- GPU memory hygiene: add autouse cleanup fixtures + _cleanup_gpu_memory()
  helpers to chemprop / fastprop ensemble and learner suites to prevent
  cumulative OOM during the slow integration run.
- Misc: clear _uint8_fallback_warned before warning-cache assertion;
  use git ls-files for tracked-checkpoint detection (avoids picking up
  untracked artifacts); explicit encoding='utf-8' on text reads;
  logger-scoped caplog.at_level; FastpropLearner val_fraction bumped to
  0.3 so the small fixture has enough rows for the validation split.

+++ tests/integration/test_parquet_flow.py
@@ -81,9 +81,9 @@ def test_parquet_schema_and_compression(parquet_flow_results):
         df = pl.read_parquet(p)
         assert 'ID' in df.columns and 'prediction' in df.columns
         assert df.schema['ID'] == pl.Utf8
-        assert df.schema['prediction'] == pl.Float64
+        assert df.schema['prediction'] == pl.Float32
         if 'uncertainty' in df.columns:
-            assert df.schema['uncertainty'] == pl.Float64
+            assert df.schema['uncertainty'] == pl.Float32
         rg = pq.ParquetFile(p).metadata.row_group(0)
         compressions = {rg.column(i).compression for i in range(rg.num_columns)}
         assert compressions == {'ZSTD'}, f'expected ZSTD, got {compressions}'
